chore: split large files (>1000 LOC) into focused modules
